feat(ui): confirm review responses as final step of the form and not with dialog window
